+++ compare-frontend/src/app/state/effects.ts
@@ -1,6 +1,6 @@
 import { Injectable } from "@angular/core";
 import { Actions, createEffect, ofType } from "@ngrx/effects";
-import { Store } from "@ngrx/store";
+import { Action, Store } from "@ngrx/store";
 import { delay, filter, map, mergeMap, switchMap, tap, withLatestFrom } from "rxjs/operators";
 import { BackendService, isExecError } from "../services/backend.service";
 import { ConfigService } from "../services/config.service";
@@ -147,28 +147,26 @@ check if more  │     └───────┬──────┘
     private openModal$ = createEffect(() => this.actions$.pipe(
         filter(action => action.type === 'Rule-set picked' || action.type === 'Step exec success'),
         withLatestFrom(this.store$.select(state => state.riesgos)),
-        map(([action, state]) => {
-            const modalData: {scenario: ScenarioName, partition: Partition, modal: ModalState}[] = [];
+        switchMap(([action, state]) => {
+            const actions: Action[] = [];
 
             const rules = getRules(state.rules);
 
             for (const [scenario, scenarioData] of Object.entries(state.scenarioData)) {
                 for (const [partition, partitionData] of Object.entries(scenarioData)) {
                     const modal = rules.modal(state, scenario as ScenarioName, partition as Partition);
-                    if (modal.args) modalData.push({scenario: scenario as ScenarioName, partition: partition as Partition, modal});
+                    if (modal.args) {  // There should be a modal ...
+                        actions.push(AppActions.openModal({scenario: scenario as ScenarioName, partition: partition as Partition, args: modal.args }));
+                    } else { // There should not be one ...
+                        if (partitionData.modal.args) { // ... but there is one:
+                            actions.push(AppActions.closeModal({scenario: scenario as ScenarioName, partition: partition as Partition}));
+                        }
+                    }
                 }
             }
 
-            return modalData;
-        }),
-            
-        switchMap(modalData => {
-            return modalData.map(m => AppActions.openModal({ 
-                scenario: m.scenario, 
-                partition: m.partition, 
-                args: m.modal.args
-            }))
-        })
+            return actions;
+        })  
     ));