+++ compare-frontend/src/app/views/mappage/mappage.component.ts
@@ -54,7 +54,6 @@ export class MappageComponent {
       for (const [partitionName, partitionData] of Object.entries(currentScenarioData)) {
         modals[partitionName] = partitionData.modal;
       }
-      // console.log("modals", modals);
       return modals;
     })
   )